feat: add footer and company carousel
