+++ components/organisms/base-header.vue
@@ -1,28 +1,8 @@
 <script setup lang="ts">
-  const LINKS = [
-    {
-      id: "linkedin",
-      iconName: "ant-design:linkedin-filled",
-      href: "https://www.linkedin.com/in/danieletortora/",
-      text: "LinkedIn",
-    },
-    {
-      id: "GitHub",
-      iconName: "ant-design:github-filled",
-      href: "https://github.com/floroz",
-      text: "GitHub",
-    },
-    {
-      id: "Twitter",
-      iconName: "ant-design:twitter-circle-filled",
-      href: "https://twitter.com/floroz87",
-      text: "Twitter",
-    },
-  ];
+  import SocialLinks from "~/components/molecules/social-links.vue";
 </script>
-
 <template>
-  <header class="max-w-4xl p-4 h-20 border-b border-b-zinc-800 m-auto">
+  <header class="header-styled border-b border-zinc-700">
     <nav>
       <ul class="flex items-center gap-x-6 w-full h-full">
         <li class="text-primary flex-grow">
@@ -42,30 +22,7 @@
             </div>
           </NuxtLink>
         </li>
-        <li
-          v-for="(link, index) in LINKS"
-          :key="link.id"
-          class="cursor-pointer text-primary hover:text-zinc-200"
-        >
-          <NuxtLink
-            :to="link.href"
-            :aria-label="link.text"
-            class="flex items-center gap-x-4"
-            target="_blank"
-          >
-            <Icon
-              class="sm:!hidden"
-              :name="link.iconName"
-              size="24"
-            ></Icon>
-            <span class="hidden sm:inline">{{ link.text }}</span>
-            <span
-              v-if="index !== LINKS.length - 1"
-              class="hidden sm:inline"
-              >/</span
-            >
-          </NuxtLink>
-        </li>
+        <SocialLinks />
       </ul>
     </nav>
   </header>
