+++ layouts/default.vue
@@ -1,8 +1,14 @@
 <script setup lang="ts">
+  import BaseFooter from "~~/components/organisms/base-footer.vue";
   import BaseHeader from "~~/components/organisms/base-header.vue";
 </script>
 
 <template>
-  <BaseHeader />
-  <slot />
+  <div class="h-auto sm:min-h-screen">
+    <BaseHeader />
+    <main>
+      <slot />
+    </main>
+    <BaseFooter />
+  </div>
 </template>
