+++ pages/index.vue
@@ -1,9 +1,11 @@
 <script setup lang="ts">
-  import HeroSection from "@/components/views/home-page/hero-section.vue";
+  import HeroSection from "~/components/views/home-page/hero-section.vue";
+  import CompanyCarousel from "~/components/views/home-page/company-carousel.vue";
 </script>
 
 <template>
   <NuxtLayout>
     <HeroSection />
+    <CompanyCarousel />
   </NuxtLayout>
 </template>
